+++ src/store/store.ts
@@ -5,25 +5,27 @@ import { persistStore } from 'redux-persist'
 import { PersistPartial } from 'redux-persist/lib/persistReducer';
 import persistingReducer from "./persistingReducer";
 
-// The redux state sync configuration
-const STATE_SYNC_CONFIG: Config = {
-    channel: 'react_todo_list_channel',
-};
-
-// All the redux middlewares
-const storeMiddlewares = [
-    createStateSyncMiddleware(STATE_SYNC_CONFIG),
-    thunk,
-];
-
-export default () => {
+export const configureStore = () => {
+    // The redux state sync configuration
+    const STATE_SYNC_CONFIG: Config = {
+        channel: 'react_todo_list_channel',
+    };
+    
+    // All the redux middlewares
+    const storeMiddlewares = [
+        createStateSyncMiddleware(STATE_SYNC_CONFIG),
+        thunk,
+    ];
+    
     // Create the redux store
     const store: Store<PersistPartial, Action<any>> & {
         dispatch: DispatchType
     } = createStore(persistingReducer, applyMiddleware(...storeMiddlewares));
-
+    
     // And it's corresponding persisto
     const persistor = persistStore(store);
 
-    return { store, persistor }
+    return { store, persistor };
 };
+
+export default () => configureStore();